added ts and moved index to src folder
