+++ packages/express/src/server.ts
@@ -1,0 +1,5 @@
+import app from './app.js'
+const port = 4040;
+app.listen(port, () => {
+  console.log(`Server is running on port ${port}`);
+});
